requirements.txt
bibtexparser==1.4.1
rapidfuzz>=3.9.0
pandas>=2.2.0
requests>=2.31.0
transformers>=4.41.0
accelerate>=0.30.0
sentencepiece>=0.2.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
unidecode>=1.3.8
